# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "轻量级音乐服务器 API 文档，支持本地音乐管理、网络歌曲、电台和歌单功能",
        "title": "MiMusic API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mimusic-org/mimusic",
            "email": "im.hanxi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.4.1"
    },
    "host": "localhost:58091",
    "basePath": "/api/v1",
    "paths": {
...
        "/playlists/{id}/convert-progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定歌单的网络歌曲→本地歌曲转换进度,无运行任务时返回 idle 状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌单转换"
                ],
                "summary": "获取歌单转换进度",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "歌单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "转换进度信息",
                        "schema": {
                            "$ref": "#/definitions/services.ConvertProgress"
                        }
                    },
                    "400": {
                        "description": "无效的歌单ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/playlists/{id}/convert-progress/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取消指定歌单的网络歌曲→本地歌曲转换任务,返回 cancelled 字段表示是否成功取消",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌单转换"
                ],
                "summary": "取消歌单转换",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "歌单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消结果,cancelled 为 true 表示成功取消",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的歌单ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/playlists/{id}/convert-to-local": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "异步将歌单内所有网络歌曲下载到本地音乐库（按 music_path/{歌单名}/{艺术家 - 标题} 命名），立即返回，可通过进度接口查询状态。同一歌单同时只允许一个任务运行。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌单转换"
                ],
                "summary": "启动歌单转换为本地歌单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "歌单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "转换任务已启动",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的歌单ID 或 歌单中没有需要转换的网络歌曲",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "该歌单已有转换任务在运行",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "启动转换失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/settings/auto-convert": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取“网络歌曲缓存完成后自动转为本地”开关的当前状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌单转换"
                ],
                "summary": "获取自动转换开关",
                "responses": {
                    "200": {
                        "description": "返回 enabled 字段表示开关状态",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "开启或关闭“网络歌曲缓存完成后自动转为本地”功能。开启后,任何 remote 类型歌曲缓存下载完成时,会异步在其所在的所有普通歌单中触发转换。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌单转换"
                ],
                "summary": "更新自动转换开关",
                "parameters": [
                    {
                        "description": "开关请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AutoConvertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回 enabled 字段表示更新后的开关状态",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "保存配置失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AutoConvertRequest": {
            "type": "object",
            "properties": {
                "enabled": {
                    "description": "是否启用自动转换",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "services.ConvertProgress": {
            "type": "object",
            "properties": {
                "converted_songs": {
                    "description": "已转换数",
                    "type": "integer"
                },
                "current_song": {
                    "description": "当前正在处理的歌曲",
                    "type": "string"
                },
                "end_time": {
                    "description": "结束时间",
                    "type": "string"
                },
                "error": {
                    "description": "致命错误",
                    "type": "string"
                },
                "errors": {
                    "description": "错误明细",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed_songs": {
                    "description": "失败数",
                    "type": "integer"
                },
                "playlist_id": {
                    "description": "歌单 ID",
                    "type": "integer"
                },
                "processed_songs": {
                    "description": "已处理数(成功+跳过+失败)",
                    "type": "integer"
                },
                "skipped_songs": {
                    "description": "跳过数(本地歌曲、已转过)",
                    "type": "integer"
                },
                "start_time": {
                    "description": "开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "当前状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/services.ConvertStatus"
                        }
                    ]
                },
                "total_songs": {
                    "description": "待转换总数",
                    "type": "integer"
                },
                "waiting": {
                    "description": "是否处于风控限速等待中",
                    "type": "boolean"
                }
            }
        },
        "services.ConvertStatus": {
            "type": "string",
            "enum": [
                "idle",
                "running",
                "completed",
                "failed",
                "cancelling",
                "cancelled"
            ],
            "x-enum-comments": {
                "ConvertStatusCancelled": "已取消",
                "ConvertStatusCancelling": "取消中",
                "ConvertStatusCompleted": "已完成",
                "ConvertStatusFailed": "失败",
                "ConvertStatusIdle": "空闲",
                "ConvertStatusRunning": "转换中"
            },
            "x-enum-descriptions": [
                "空闲",
                "转换中",
                "已完成",
                "失败",
                "取消中",
                "已取消"
            ],
            "x-enum-varnames": [
                "ConvertStatusIdle",
                "ConvertStatusRunning",
                "ConvertStatusCompleted",
                "ConvertStatusFailed",
                "ConvertStatusCancelling",
                "ConvertStatusCancelled"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\" 进行认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}